Implement create new house functionality
Create new house is now connected to firestore and actually creates a new House.
After a successful house creation the user is moved to the house activity.
Splash screen moves the user to LoginActivity if he is not logged in, to ChooseHouseActivity if he is logged in but doesn't have a house, or to HouseActivity if he is logged in and has a house.

Along with this, we add FirestoreJob class which allows us to communicate with UI via LiveData when firestore jobs are done.