Redirect user to correct activity after sign in
After a user signs in, redirect him to the correct activity:
ChooseHouseActivity if he doesn't have a house or to HouseActivity if he
already has a house.
Until now, user was redirected only to ChooseHouseActivity.

Also change the name of LoginActivity.java to SignInActivity to create
a uniform terminology with FirebaseAuthUI.