Add users collection, add repositories
Added users collection so now other users data (and not only current user)
can be queried.

For now, the user document is updated in the following cases:
- After sign-in, a new user document is created (if the user doesn't
  already exist in the users collection).
- After new house creation, the user role is updated.
- After joining to a house, the user role is updated.

Also added repositories (for House and User for now) that will be the
single source of truth for house and user related queries.